+++ spec/app/models/name.rb
@@ -8,6 +8,7 @@ class Name
   field :first_name, type: String
   field :last_name, type: String
   field :parent_title, type: String
+  field :middle, type: String
 
   embeds_many :translations, validate: false
   embeds_one :language, as: :translatable, validate: false